Prevent overwriting of HTML when combined with JSX and/or Handlebars (#24)
* Fix: Allow to mix `html` and framework examples

We have `html` and also React examples. This commit fixes the problem that the frameworks have overwritten the HTML example.

* Prevent overwriting of HTML when combined with JSX and/or Handlebars

If an example contained an HTML as well as a JSX/Angular/Handlebars block, the HTML content was being inadvertently overwritten by the latter block.

This fixes that behavior by using individual DOM elements to bind each JSX, Angular, or Handlebars block.