Replace form tag to prevent page refresh when hitting enter
With a form tag, the default browser behavior when hitting ENTER while focused on the search input is to submit the page to the URL specified by the 'action' attribute. Since there's no such attribute on that form tag, it defaults to submitting to the same page. This causes a page refresh which clears the search input, making it seem as if the search doesn't work.

Fixed by simply replacing the form tag with a <div> to avoid the default browser behavior.

Fixes #9