For external example code blocks put closing backticks on their own line
If the external example file didn't have an empty newline at the end of the file, the closing triple backticks (```) in the resulting code block weren't on their own line. This caused later parsing of that code block to fail.

For example, given this external example reference:

    ```example:files/style.css
    ```

and this files/style.css:

    p {
        color: red;
    }EOF

where EOF was the end of the file (ie. no ending newline).

The resulting rendered block was:

    ```example.css
    p {
        color: red
    }```

Note how ``` does not appear on its own line.

This commit fixes that so that ``` always appears on its own line:

    ```example.css
    p {
        color: red
    }
    ```

Fixes #39